refactor: UnauthenticatedRoute.js -> UnauthenticatedRoutes.tsx
